requirements.txt
numpy==1.14.5
tensorflow-gpu=1.9.0
pillow
pynbody
h5py
sklearn
scikit-image
pyyaml
-e .
-e tfnntools
matplotlib
jupyter
moviepy
requests
astropy
lenstools